port: backport UnrealMCP plugin to UE4.27 + VS2019
- Fix 6 API incompatibilities (IWYUSupport, BuildSettings V5->V2, EditorActorSubsystem, FMath::Max, BufferSize rename, drop BlueprintEditorLibrary)
- Verified runtime on UE4.27: actor ops + programmatic blueprint create/compile/spawn all pass
- Rewrite README for this project; add WorkLog 2026-07-01

+++ MCPGameProject/Plugins/UnrealMCP/Source/UnrealMCP/Private/Commands/UnrealMCPUMGCommands.cpp
@@ -393,7 +393,8 @@ TSharedPtr<FJsonObject> FUnrealMCPUMGCommands::HandleBindWidgetEvent(const TShar
 		float MaxHeight = 0.0f;
 		for (UEdGraphNode* Node : EventGraph->Nodes)
 		{
-			MaxHeight = FMath::Max(MaxHeight, Node->NodePosY);
+			// UE4.27: NodePosY is int32; cast to float so FMath::Max template arg is unambiguous.
+			MaxHeight = FMath::Max(MaxHeight, (float)Node->NodePosY);
 		}
 		
 		const FVector2D NodePos(200, MaxHeight + 200);
